requirements.txt
matplotlib==3.9.2
numpy==2.0.2
pillow==11.0.0
python-dotenv==1.0.1